+++ CHANGELOG.rst
@@ -65,7 +65,8 @@ Version 2021-dev
 -  Add a molden parser to xtp (#547)
 -  Correct self-energy frequency derivative (#550)
 -  fix multiline orca keyword (#552)
--  Contour Deformation Approach (CDA) for self-energy implmented (#4)
+-  Contour Deformation Approach (CDA) for self-energy implemented (#4)
+-  BSE with perturbative dynamical screening (#3)
 
 Version 1.6.3 (released XX.08.20)
 =================================